dummy redirect to secure dashboard from /login
