--- templates/auth/login_modal.html
@@ -1,33 +1,0 @@
-<div class="flex justify-between items-center mb-4">
-  <h2 class="text-xl font-bold">Login</h2>
-  <button onclick="document.getElementById('loginModal').classList.add('hidden')"
-    class="text-gray-500 hover:text-gray-700">
-    ✕
-  </button>
-</div>
-<form hx-post="/token" hx-target="#loginModal">
-  <div class="mb-4">
-    <label class="block text-gray-700 text-sm font-bold mb-2" for="username">
-      Username
-    </label>
-    <input
-      class="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-      id="username" name="username" type="text" required>
-  </div>
-  <div class="mb-6">
-    <label class="block text-gray-700 text-sm font-bold mb-2" for="password">
-      Password
-    </label>
-    <input
-      class="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 mb-3 leading-tight focus:outline-none focus:shadow-outline"
-      id="password" name="password" type="password" required>
-  </div>
-  <div class="flex items-center justify-between">
-    <button
-      class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
-      type="submit">
-      Sign In
-    </button>
-  </div>
-</form>
-</div>
