+++ templates/public/public_dashboard.html
@@ -253,7 +253,7 @@
         <li><a href="/">Home</a></li>
         <li><a href="/about">About</a></li>
         <li><a href="/contact">Contact</a></li>
-        <li><a href="/dashboard" class="login-btn">Login</a></li>
+        <li><a href="/login" class="login-btn">Login</a></li>
       </ul>
     </div>
   </nav>
